+++ src/extension-test/unit/cursor-doc/indent-test.ts
@@ -142,6 +142,30 @@ describe('indent', () => {
         ).toEqual(2);
       });
     });
+
+    describe('deftype', () => {
+      it('calculates indents for cursor on the new line of a method implementation', () => {
+        const doc = docFromTextNotation(`
+(deftype MyType [arg1 arg2]
+  IMyProto
+  (method1 [this]
+|(print "hello")))`);
+
+        expect(
+          indent.getIndent(
+            doc.model,
+            textAndSelection(doc)[1][0],
+            mkConfig({
+              deftype: [
+                ['block', 2],
+                ['inner', 1],
+              ],
+              '#"^def(?!ault)(?!late)(?!er)"': [['inner', 0]],
+            })
+          )
+        ).toEqual(4);
+      });
+    });
   });
 
   describe('collectIndents', () => {